requirements.txt
selenium==3.0.1
pyyaml==3.12

